Redesigned the TTP classes and created new tests

+++ src/main/java/com/moo/ttp/operators/BitFlip.java
@@ -7,7 +7,7 @@ public class BitFlip {
 	
 	
 	public static void mutate(Boolean[] b) {
-		double prob = 1 / b.length;
+		double prob = 1 / (double) b.length;
 		for (int i = 0; i < b.length; i++) {
 			if (Rnd.rndDouble() < prob) b[i] = !b[i];
 		}
